Move CallTimings around
