+++ src/config.rs
@@ -1,5 +1,4 @@
 use std::fmt;
-use std::sync::Arc;
 use std::time::Duration;
 
 use hoot::client::flow::RedirectAuthHeaders;
@@ -7,8 +6,7 @@ use http::Uri;
 
 use crate::middleware::MiddlewareChain;
 use crate::resolver::IpFamily;
-use crate::transport::time::{Instant, NextTimeout};
-use crate::{Proxy, TimeoutReason};
+use crate::Proxy;
 
 #[cfg(feature = "_tls")]
 use crate::tls::TlsConfig;
@@ -323,161 +321,5 @@ impl fmt::Debug for Timeouts {
             .field("recv_response", &self.recv_response)
             .field("recv_body", &self.recv_body)
             .finish()
-    }
-}
-
-#[derive(Debug, Default)]
-pub(crate) struct CallTimings {
-    pub timeouts: Timeouts,
-    pub current_time: CurrentTime,
-    pub time_global_start: Option<Instant>,
-    pub time_call_start: Option<Instant>,
-    pub time_resolve: Option<Instant>,
-    pub time_connect: Option<Instant>,
-    pub time_send_request: Option<Instant>,
-    pub time_send_body: Option<Instant>,
-    pub time_await_100: Option<Instant>,
-    pub time_recv_response: Option<Instant>,
-    pub time_recv_body: Option<Instant>,
-}
-
-#[derive(Clone)]
-pub(crate) struct CurrentTime(Arc<dyn Fn() -> Instant + Send + Sync + 'static>);
-
-impl CurrentTime {
-    pub(crate) fn now(&self) -> Instant {
-        self.0()
-    }
-}
-
-impl CallTimings {
-    pub(crate) fn now(&self) -> Instant {
-        self.current_time.now()
-    }
-
-    pub(crate) fn record_timeout(&mut self, reason: TimeoutReason) {
-        match reason {
-            TimeoutReason::Global => {
-                let now = self.now();
-                if self.time_global_start.is_none() {
-                    self.time_global_start = Some(now);
-                }
-                self.time_call_start = Some(now);
-            }
-            TimeoutReason::Resolver => {
-                self.time_resolve = Some(self.now());
-            }
-            TimeoutReason::OpenConnection => {
-                self.time_connect = Some(self.now());
-            }
-            TimeoutReason::SendRequest => {
-                self.time_send_request = Some(self.now());
-            }
-            TimeoutReason::SendBody => {
-                self.time_send_body = Some(self.now());
-            }
-            TimeoutReason::Await100 => {
-                self.time_await_100 = Some(self.now());
-            }
-            TimeoutReason::RecvResponse => {
-                self.time_recv_response = Some(self.now());
-            }
-            TimeoutReason::RecvBody => {
-                self.time_recv_body = Some(self.now());
-            }
-        }
-    }
-
-    pub(crate) fn next_timeout(&self, reason: TimeoutReason) -> NextTimeout {
-        // self.time_xxx unwraps() below are OK. If the unwrap fails, we have a state
-        // bug where we progressed to a certain state without setting the corresponding time.
-        let timeouts = &self.timeouts;
-
-        let expire_at = match reason {
-            TimeoutReason::Global => timeouts
-                .global
-                .map(|t| self.time_global_start.unwrap() + t.into()),
-            TimeoutReason::Resolver => timeouts
-                .resolve
-                .map(|t| self.time_call_start.unwrap() + t.into()),
-            TimeoutReason::OpenConnection => timeouts
-                .connect
-                .map(|t| self.time_resolve.unwrap() + t.into()),
-            TimeoutReason::SendRequest => timeouts
-                .send_request
-                .map(|t| self.time_connect.unwrap() + t.into()),
-            TimeoutReason::SendBody => timeouts
-                .send_body
-                .map(|t| self.time_send_request.unwrap() + t.into()),
-            TimeoutReason::Await100 => timeouts
-                .await_100
-                .map(|t| self.time_send_request.unwrap() + t.into()),
-            TimeoutReason::RecvResponse => timeouts.recv_response.map(|t| {
-                // The fallback order is important. See state diagram in hoot.
-                self.time_send_body
-                    .or(self.time_await_100)
-                    .or(self.time_send_request)
-                    .unwrap()
-                    + t.into()
-            }),
-            TimeoutReason::RecvBody => timeouts
-                .recv_body
-                .map(|t| self.time_recv_response.unwrap() + t.into()),
-        }
-        .unwrap_or(Instant::NotHappening);
-
-        let global_at = self.global_timeout();
-
-        let (at, reason) = if global_at < expire_at {
-            (global_at, TimeoutReason::Global)
-        } else {
-            (expire_at, reason)
-        };
-
-        let after = at.duration_since(self.now());
-
-        NextTimeout { after, reason }
-    }
-
-    fn global_timeout(&self) -> Instant {
-        let global_start = self.time_global_start.unwrap();
-        let call_start = self.time_call_start.unwrap();
-
-        let global_at = global_start
-            + self
-                .timeouts
-                .global
-                .map(|t| t.into())
-                .unwrap_or(crate::transport::time::Duration::NotHappening);
-
-        let call_at = call_start
-            + self
-                .timeouts
-                .per_call
-                .map(|t| t.into())
-                .unwrap_or(crate::transport::time::Duration::NotHappening);
-
-        global_at.min(call_at)
-    }
-
-    pub(crate) fn new_call(self) -> CallTimings {
-        CallTimings {
-            timeouts: self.timeouts,
-            time_global_start: self.time_global_start,
-            current_time: self.current_time,
-            ..Default::default()
-        }
-    }
-}
-
-impl fmt::Debug for CurrentTime {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_tuple("CurrentTime").finish()
-    }
-}
-
-impl Default for CurrentTime {
-    fn default() -> Self {
-        Self(Arc::new(Instant::now))
     }
 }
